feat(HTD-52): move mdx blog method to blog folder
